+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `"reply"` argument type for getting in replies.
+- Editing posts.
 
 ## 1.7.3 - 2024-11-08